Split messages out into own file
